Extract getDataSourceUrl api, add url to edit form
